refactor: TossRestClientConfig 코드 수정
