--- src/main/java/roomescape/payment/dto/PaymentProviderInfo.java
@@ -1,9 +1,0 @@
-package roomescape.payment.dto;
-
-public record PaymentProviderInfo(
-        String secretKey,
-        String baseUrl,
-        int connectTimeout,
-        int readTimeout
-) {
-}
