+++ src/main/resources/application.yml
@@ -23,9 +23,8 @@ security:
       expire-length: 3600000
 
 payment:
-  properties:
-    toss:
-      secret-key: test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6
-      base-url: https://api.tosspayments.com
-      connect-timeout: 5000
-      read-timeout: 30000
+  toss:
+    secret-key: test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6
+    base-url: https://api.tosspayments.com
+    connect-timeout: 5000
+    read-timeout: 30000
